Finish website layout
This commit adds 4 images to information section.
Photos wrap, align and justify to the center.
README.md contains links to all images in the
sources section. Changes in index.html in order to
have images in right order. This concludes website
layout. Now, it only needs appropriate text in place
of each placeholder.